Added error handling to the endpoints on both the explorer and creature sides. Also added testing functions to the creature side.

+++ FastAPI_practice/fastapi_app/src/data/creature.py
@@ -1,5 +1,6 @@
-from .init import conn, curs
+from .init import conn, curs, IntegrityError
 from model.creature import Creature
+from error import Missing, Duplicate
 
 curs.execute("""create table if not exists creature(
     name text primary key,
@@ -14,32 +15,48 @@ def row_to_model(row: tuple) -> Creature:
     return Creature(name, description, country, area, aka)
 
 
-def model_to_dict(creature: Creature) -> dict:
-    return creature.dict()
+def model_to_dict(creature: Creature) -> dict | None:
+    return creature.dict() if creature else None
 
 
 def get_one(name: str) -> Creature:
     qry = "select * from creature where name:=name"
     params = {"name": name}
     curs.execute(qry, params)
-    return row_to_model(curs.fetchone())
+    row = curs.fetchone()
+    if row:
+        return row_to_model(row)
+    else:
+        raise Missing(msg=f"Creature {name} not found.")
 
 
-def get_all(name: str) -> list[Creature]:
+def get_all() -> list[Creature]:
     qry = "select * from creature"
     curs.execute(qry)
     rows = list(curs.fetchall())
     return [row_to_model(row) for row in rows]
 
 
-def create(creature: Creature):
+def create(creature: Creature) -> Creature:
+    if not creature:
+        return None
+
     qry = """insert into creature values (:name, :description, :country, :area, :aka)"""
     params = model_to_dict(creature)
-    curs.execute(qry, params)
+
+    try:
+        curs.execute(qry, params)
+        conn.commit()
+    except IntegrityError:
+        raise Duplicate(msg=f"Creature {creature.name} already exists.")
+
     return get_one(creature.name)
 
 
-def modify(creature: Creature):
+def modify(name: str, creature: Creature) -> Creature | None:
+    if not (name and creature):
+        return None
+
     qry = """update creature set
              country = :country,
              description = :description,
@@ -47,18 +64,29 @@ def modify(creature: Creature):
              aka = :aka
              where name = :name_orig"""
     params = model_to_dict(creature)
-    params["name_orig"] = creature.name
-    _ = curs.execute(qry, params)
+    params["name_orig"] = name
+    curs.execute(qry, params)
 
-    return get_one(creature.name)
+    if curs.rowcount == 1:
+        return get_one(creature.name)
+    else:
+        raise Missing(msg=f"Creature {name} not found.")
 
 
-def replace(creature: Creature):
+def replace(name: str, creature: Creature):
     return creature
 
 
-def delete(creature: Creature) -> bool:
+def delete(name: str):
+    if not name:
+        return False
+
     qry = "delete from creature where name = :name"
-    params = {"name": creature.name}
+    params = {"name": name}
     res = curs.execute(qry, params)
-    return bool(res)
+
+    if curs.rowcount == 1:
+        conn.commit()
+        return bool(res)
+    else:
+        raise Missing(msg=f"Creature {name} not found.")
