+++ FastAPI_practice/fastapi_app/src/data/explorer.py
@@ -1,5 +1,6 @@
-from .init import conn, curs
+from .init import conn, curs, IntegrityError
 from model.explorer import Explorer
+from error import Missing, Duplicate
 
 curs.execute("""create table if not exists explorer(
                 name text primary key,
@@ -8,9 +9,9 @@ curs.execute("""create table if not exists explorer(
 
 
 def row_to_model(row: tuple) -> Explorer:
-    print("________________________________")
-    print("this is the row", row)
-    print("________________________________")
+    # print("________________________________")
+    # print("this is the row", row)
+    # print("________________________________")
     return Explorer(name=row[0], country=row[1], description=row[2])
 
 
@@ -22,13 +23,14 @@ def get_one(name: str) -> Explorer:
     qry = "select * from explorer where name = ?"
     params = (name,)
     curs.execute(qry, params)
-    print("________________________________")
-    print("executing sql inside get_one", qry, params)
-    print("________________________________")
-    return row_to_model(curs.fetchone())
+    row = curs.fetchone()
+    if row:
+        return row_to_model(row)
+    else:
+        raise Missing(msg=f"Explorer {name} not found.")
 
 
-def get_all(name: str | None) -> list[Explorer]:
+def get_all() -> list[Explorer]:
     qry = "select * from explorer"
     curs.execute(qry)
     rows = list(curs.fetchall())
@@ -36,35 +38,57 @@ def get_all(name: str | None) -> list[Explorer]:
 
 
 def create(explorer: Explorer):
-    qry = """INSERT INTO explorer(name, country, description)
-             VALUES(?, ?, ?)"""
-    # print("this is the explorer object", explorer)
-    params = tuple(model_to_dict(explorer).values())
-    # print("Executing SQL inside create: ", qry, params)
-    _ = curs.execute(qry, params)
-    conn.commit()
+    if not explorer:
+        return None
+
+    qry = """INSERT INTO explorer(name, description, country)
+             VALUES(:name, :country, :description)"""
+    params = model_to_dict(explorer)
+
+    try:
+        curs.execute(qry, params)
+        conn.commit()
+    except IntegrityError:
+        raise Duplicate(msg=f"Explorer {explorer.name} already exists.")
+
     return get_one(explorer.name)
-    # return {"name": explorer.name, "success": True}
 
 
-def modify(name: str, explorer: Explorer) -> Explorer:
+# NOTE: Should the name string be the one to be passed to the name_orig parameter since that
+# is what is being read to select the row and update it with the body parameters
+def modify(name: str, explorer: Explorer) -> Explorer | None:
+    if not (name and explorer):
+        return None
+
     qry = """update explorer set
              name = :name,
              country = :country,
              description = :description,
              where name = :name_orig"""
     params = model_to_dict(explorer)
-    params["name_orig"] = explorer.name
-    _ = curs.execute(qry, params)
-    return get_one(explorer.name)
+    params["name_orig"] = name
+    curs.execute(qry, params)
+
+    if curs.rowcount == 1:
+        return get_one(explorer.name)
+    else:
+        raise Missing(msg=f"Explorer {name} not found.")
 
 
 def replace(name: str, explorer: Explorer):
     return explorer
 
 
-def delete(explorer: Explorer) -> bool:
+def delete(name: str):
+    if not name:
+        return False
+
     qry = "delete from explorer where name = :name"
-    params = {"name": explorer.name}
+    params = {"name": name}
     res = curs.execute(qry, params)
-    return bool(res)
+
+    if curs.rowcount == 1:
+        conn.commit()
+        return bool(res)
+    else:
+        raise Missing(msg=f"Explorer {name} not found.")
